bug: assembly 启动失败[1.5.4] #4475 解决默认datasource问题
